Move more attributes to AstroData
